requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
huggingface_hub>=0.15.0
Pillow>=10.0.0
streamlit>=1.26.0

# Model-specific requirements
einops>=0.6.0
timm>=0.9.0

# Additional utilities
SpeechRecognition>=3.10.0
fpdf>=1.7.2
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
nltk>=3.8.1
wordcloud>=1.9.0

